+++ Wire-iOS/Sources/UserInterface/Calling/CallDegradationViewController/CallDegradationController.swift
@@ -82,47 +82,5 @@ final class CallDegradationController: UIViewController {
         Log.calling.debug("Presenting alert about degraded call")
         targetViewController?.present(alertViewController, animated: !ProcessInfo.processInfo.isRunningTests)
     }
-    
 }
 
-extension UIAlertController {
-    
-    static func degradedCall(degradedUser: ZMUser?, callEnded: Bool = false, confirmationBlock: ((_ continueDegradedCall: Bool) -> Void)? = nil) -> UIAlertController {
-        
-        // Choose localization prefix
-        let prefix = callEnded
-            ? "call.degraded.ended.alert"
-            : "call.degraded.alert"
-       
-        // Set message
-        var message = "\(prefix).message"
-        
-        switch degradedUser {
-        case .some(let user) where user.isSelfUser:
-            message = "\(message).self".localized
-        case .some(let user):
-            message = "\(message).user".localized(args: user.name ?? "")
-        default:
-            message = "\(message).unknown".localized
-        }
-        
-        // Create controller
-        let controller = UIAlertController(title: "\(prefix).title".localized, message: message, preferredStyle: .alert)
-        
-        // Add actions
-        if let confirmationBlock = confirmationBlock {
-            controller.addAction(UIAlertAction(title: "general.cancel".localized, style: .cancel) { (action) in
-                confirmationBlock(false)
-            })
-            
-            controller.addAction(UIAlertAction(title: "call.degraded.alert.action.continue".localized, style: .default) { (action) in
-                confirmationBlock(true)
-            })
-        } else {
-            controller.addAction(UIAlertAction(title: "general.ok".localized, style: .default))
-        }
-        
-        return controller
-    }
-    
-}
